Refactored namespace for the main server deliverable.
Now everything should be under net.krotscheck.kangaroo
