+++ kangaroo-server/pom.xml
@@ -67,6 +67,11 @@
     </dependency>
     <dependency>
       <groupId>net.krotscheck</groupId>
+      <artifactId>kangaroo-servlet-admin</artifactId>
+      <version>${project.parent.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>net.krotscheck</groupId>
       <artifactId>kangaroo-test</artifactId>
       <version>${project.parent.version}</version>
       <scope>test</scope>
